backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
pydantic==2.11.2
pydantic-settings==2.9.1
email-validator==2.2.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-multipart==0.0.20
azure-storage-blob==12.25.1
azure-ai-documentintelligence==1.0.2
langgraph==0.6.7
httpx==0.28.1
